Move Step component to root
